+++ src/main/java/org/example/odiya/notification/service/event/HurryUpEvent.java
@@ -1,4 +1,4 @@
-package org.example.odiya.notification.dto.request;
+package org.example.odiya.notification.service.event;
 
 import lombok.Getter;
 import org.example.odiya.mate.domain.Mate;
@@ -6,12 +6,12 @@ import org.example.odiya.notification.domain.Notification;
 import org.springframework.context.ApplicationEvent;
 
 @Getter
-public class HurryUpRequest extends ApplicationEvent {
+public class HurryUpEvent extends ApplicationEvent {
 
     private final Mate mate;
     private final Notification notification;
 
-    public HurryUpRequest(Object object, Mate mate, Notification notification) {
+    public HurryUpEvent(Object object, Mate mate, Notification notification) {
         super(object);
         this.mate = mate;
         this.notification = notification;
