+++ src/main/java/org/example/odiya/notification/service/event/NoticeEvent.java
@@ -1,15 +1,15 @@
-package org.example.odiya.notification.dto.request;
+package org.example.odiya.notification.service.event;
 
 import lombok.Getter;
 import org.example.odiya.notification.domain.message.GroupMessage;
 import org.springframework.context.ApplicationEvent;
 
 @Getter
-public class NoticeRequest extends ApplicationEvent {
+public class NoticeEvent extends ApplicationEvent {
 
     private final GroupMessage groupMessage;
 
-    public NoticeRequest(Object object, GroupMessage groupMessage) {
+    public NoticeEvent(Object object, GroupMessage groupMessage) {
         super(object);
         this.groupMessage = groupMessage;
     }
