+++ src/main/java/org/example/odiya/notification/service/event/PushEvent.java
@@ -1,4 +1,4 @@
-package org.example.odiya.notification.dto.request;
+package org.example.odiya.notification.service.event;
 
 import lombok.Getter;
 import org.example.odiya.notification.domain.Notification;
@@ -6,12 +6,12 @@ import org.example.odiya.notification.domain.message.GroupMessage;
 import org.springframework.context.ApplicationEvent;
 
 @Getter
-public class PushRequest extends ApplicationEvent {
+public class PushEvent extends ApplicationEvent {
 
     private final Notification notification;
     private final GroupMessage groupMessage;
 
-    public PushRequest(Object object, Notification notification) {
+    public PushEvent(Object object, Notification notification) {
         super(object);
         this.notification = notification;
         this.groupMessage = GroupMessage.createGlobalNotice(notification);
