+++ src/main/java/org/example/odiya/notification/service/event/SubscribeEvent.java
@@ -1,4 +1,4 @@
-package org.example.odiya.notification.dto.request;
+package org.example.odiya.notification.service.event;
 
 import lombok.Getter;
 import org.example.odiya.member.domain.DeviceToken;
@@ -6,13 +6,13 @@ import org.example.odiya.notification.domain.FcmTopic;
 import org.springframework.context.ApplicationEvent;
 
 @Getter
-public class SubscribeRequest extends ApplicationEvent {
+public class SubscribeEvent extends ApplicationEvent {
 
     private final DeviceToken deviceToken;
     private final FcmTopic fcmTopic;
 
 
-    public SubscribeRequest(Object object, DeviceToken deviceToken, FcmTopic fcmTopic) {
+    public SubscribeEvent(Object object, DeviceToken deviceToken, FcmTopic fcmTopic) {
         super(object);
         this.deviceToken = deviceToken;
         this.fcmTopic = fcmTopic;
